Remove taglib_sys::set_strings_unicode() and wrappers
Strings inside TagLib are considered to be UTF-8 (even if they are
stored internally as UTF-16 because of reasons), and disabling that
will make everything fall back to ISO-8859-1, which is utterly
broken in 2015.

Since Rust considers strings to be UTF-8, let's paper over this whole
mess, and remove a "break my application" API. If you fall into a
wormhole to 1997 and feel the need to use Latin-1 for your strings, you
can very likely just add it back yourself.