Store tag and property content directly in struct
Previously a handle to `File` was kept around in `Tag` and `AudioProperties`
so that the raw file pointer wouldn't be dropped while it was being used.

This made it difficult to just pass the tag around as a user of the API,
due to having to keep the `file` object alive as well.

For example, this would result in a lifetime error because the file object
does not live long enough:

```
taglib::File::new(path).and_then(|f| f.tag())
```

In this change the `Tag` and `AudioProperties` are created when request from
the file, E.g. `file.tag()` and there is no raw pointers being passed around.

If the user wishes to save new tags they pass the tag struct in to the `save`
function directly instead (`AudioProperties` can't be changed).